Move game_service to own folder.
