+++ lib/widgets/game_screen/question_answer_panel.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:auto_size_text/auto_size_text.dart';
-import 'package:bratacha/services/i_game_service.dart';
+import 'package:bratacha/services/game_service/i_game_service.dart';
 import 'package:bratacha/widgets/common/flag.dart';
 import 'package:bratacha/widgets/game_screen/answers_cubit.dart';
 import 'package:bratacha/widgets/game_screen/question_cubit.dart';
